+++ src/neptune/metadata_containers/utils.py
@@ -15,12 +15,10 @@
 #
 
 __all__ = [
-    "DATE_FORMAT",
     "parse_dates",
     "prepare_nql_query",
 ]
 
-from datetime import datetime
 from typing import (
     Generator,
     Iterable,
@@ -47,9 +45,8 @@ from neptune.internal.backends.nql import (
     NQLQueryAttribute,
     RawNQLQuery,
 )
+from neptune.internal.utils.iso_dates import parse_iso_date
 from neptune.internal.utils.run_state import RunState
-
-DATE_FORMAT: str = "%Y-%m-%dT%H:%M:%S.%fZ"
 
 
 def prepare_nql_query(
@@ -154,7 +151,7 @@ def _parse_entry(entry: LeaderboardEntry) -> LeaderboardEntry:
                         attribute.type,
                         {
                             **attribute.properties,
-                            "value": datetime.strptime(attribute.properties["value"], DATE_FORMAT),
+                            "value": parse_iso_date(attribute.properties["value"]),
                         },
                     )
                     if attribute.type == AttributeType.DATETIME
@@ -163,9 +160,6 @@ def _parse_entry(entry: LeaderboardEntry) -> LeaderboardEntry:
                 for attribute in entry.attributes
             ],
         )
-    except TypeError:
-        # date is already in the right format
-        return entry
     except ValueError:
         # the parsing format is incorrect
         warn_once(