+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Improve dependency installation checking ([#1670](https://github.com/neptune-ai/neptune-client/pull/1670))
 - Cache dependencies check ([#1675](https://github.com/neptune-ai/neptune-client/pull/1675))
 - Remove deprecation warning in `fetch_runs_table()` ([#1686](https://github.com/neptune-ai/neptune-client/pull/1686))
+- Make sleep dur in some e2e tests consistent (#[1700](https://github.com/neptune-ai/neptune-client/pull/1700))
 
 ### Changes
 - Minor updates to made package compatible with Python 3.11 and 3.12 ([#1678](https://github.com/neptune-ai/neptune-client/pull/1678))