+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Changes
 - Improve dependency installation checking ([#1670](https://github.com/neptune-ai/neptune-client/pull/1670))
 - Cache dependencies check ([#1675](https://github.com/neptune-ai/neptune-client/pull/1675))
+- Improve datetime parsing in `fetch_*_table()` ([#1680](https://github.com/neptune-ai/neptune-client/pull/1680))
 - Remove deprecation warning in `fetch_runs_table()` ([#1686](https://github.com/neptune-ai/neptune-client/pull/1686))
 
 ### Changes